fix: systemd on user level (#57)
We were masking some systemd units in two parts:

1. Due the similar nature, the WSL images inherited masking of some
systemd units from the init containers.

2. There is also a list of systemd units are recommended to be masked
by the official WSL documentation with introduction of the
new packaging format and installation format.

Dropping the first part lets systemd work in user level too.
So, instead of hunting down on which units breaks user level support
when it is masked, it is better to only rely on the official WSL
documentation, which built from the WSL specific reports.

Signed-off-by: Elkhan Mammadli <elkhan.mammadli@protonmail.com>